vector-service/requirements.txt
# vector-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
numpy==1.25.2
pydantic==2.4.2
httpx==0.25.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
python-multipart==0.0.6